Split AlertRulesPage out into a separate file
