+++ web/src/components/alerting.tsx
@@ -1,286 +1,35 @@
-import {
-  Alert,
-  AlertSeverity,
-  AlertStates,
-  ListPageFilter,
-  PrometheusAlert,
-  RowFilter,
-  RowProps,
-  Rule,
-  TableColumn,
-  useActiveNamespace,
-  useListPageFilter,
-  useResolvedExtensions,
-  VirtualizedTable,
-} from '@openshift-console/dynamic-plugin-sdk';
-import { sortable } from '@patternfly/react-table';
+import { useActiveNamespace, useResolvedExtensions } from '@openshift-console/dynamic-plugin-sdk';
 import classNames from 'classnames';
-import * as _ from 'lodash-es';
 import * as React from 'react';
-import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { Link, Redirect, Route, RouteComponentProps, Switch } from 'react-router-dom';
 
-import { withFallback } from './console/console-shared/error/error-boundary';
 import { AlertingRulesSourceExtension, isAlertingRulesSource } from './console/extensions/alerts';
 import { getAllQueryArguments } from './console/utils/router';
-import { EmptyBox } from './console/utils/status-box';
 
 import MonitoringDashboardsPage from './dashboards';
 import { QueryBrowserPage } from './metrics';
 import { CreateSilence, EditSilence } from './silence-form';
 import { TargetsUI } from './targets';
-import { Alerts, AlertSource } from './types';
-import {
-  alertingRuleStateOrder,
-  alertSeverityOrder,
-  fuzzyCaseInsensitive,
-  RuleResource,
-} from './utils';
 
 import './_monitoring.scss';
 import {
   getAlertRulesUrl,
   getAlertsUrl,
   getIncidentsUrl,
-  getObserveState,
-  getRuleUrl,
   getSilencesUrl,
   usePerspective,
 } from './hooks/usePerspective';
-import {
-  alertingRuleSource,
-  AlertStateIcon,
-  getAdditionalSources,
-  getAlertStateKey,
-  MonitoringResourceIcon,
-  Severity,
-  SilencesNotLoadedWarning,
-} from './alerting/AlertUtils';
 import AlertsPage from './alerting/AlertsPage';
 import AlertsDetailsPage from './alerting/AlertsDetailPage';
 import { useRulesAlertsPoller } from './hooks/useRulesAlertsPoller';
 import { useSilencesPoller } from './hooks/useSilencesPoller';
-import { MonitoringState } from '../reducers/observe';
 import SilencesPage from './alerting/SilencesPage';
 import SilencesDetailsPage from './alerting/SilencesDetailPage';
 import { useFeatures } from './hooks/useFeatures';
 import AlertRulesDetailsPage from './alerting/AlertRulesDetailsPage';
-
-const StateCounts: React.FC<{ alerts: PrometheusAlert[] }> = ({ alerts }) => {
-  const { t } = useTranslation(process.env.I18N_NAMESPACE);
-
-  const counts = _.countBy(alerts, 'state');
-  const states = [AlertStates.Firing, AlertStates.Pending, AlertStates.Silenced].filter(
-    (s) => counts[s] > 0,
-  );
-
-  return (
-    <>
-      {states.map((s) => (
-        <div className="monitoring-icon-wrap" key={s}>
-          <AlertStateIcon state={s} /> {counts[s]} {getAlertStateKey(s, t)}
-        </div>
-      ))}
-    </>
-  );
-};
-
-export const severityRowFilter = (t): RowFilter => ({
-  filter: (filter, alert: Alert) =>
-    filter.selected?.includes(alert.labels?.severity) || _.isEmpty(filter.selected),
-  filterGroupName: t('Severity'),
-  items: [
-    { id: AlertSeverity.Critical, title: t('Critical') },
-    { id: AlertSeverity.Warning, title: t('Warning') },
-    { id: AlertSeverity.Info, title: t('Info') },
-    { id: AlertSeverity.None, title: t('None') },
-  ],
-  reducer: ({ labels }: Alert | Rule) => labels?.severity,
-  type: 'alert-severity',
-});
-
-const ruleHasAlertState = (rule: Rule, state: AlertStates): boolean =>
-  state === AlertStates.NotFiring ? _.isEmpty(rule.alerts) : _.some(rule.alerts, { state });
-
-const ruleAlertStateFilter = (filter, rule: Rule) =>
-  (filter.selected?.includes(AlertStates.NotFiring) && _.isEmpty(rule.alerts)) ||
-  _.some(rule.alerts, (a) => filter.selected?.includes(a.state)) ||
-  _.isEmpty(filter.selected);
-
-export const alertStateFilter = (t): RowFilter => ({
-  filter: ruleAlertStateFilter,
-  filterGroupName: t('Alert State'),
-  isMatch: ruleHasAlertState,
-  items: [
-    { id: AlertStates.Firing, title: t('Firing') },
-    { id: AlertStates.Pending, title: t('Pending') },
-    { id: AlertStates.Silenced, title: t('Silenced') },
-    { id: AlertStates.NotFiring, title: t('Not Firing') },
-  ],
-  type: 'alerting-rule-has-alert-state',
-});
-
-const tableRuleClasses = [
-  'pf-u-w-50 pf-u-w-33-on-sm', // Name
-  'pf-m-hidden pf-m-visible-on-sm', // Severity
-  '', // Alert state
-  'pf-m-hidden pf-m-visible-on-sm', // Source
-];
-
-const RuleTableRow: React.FC<RowProps<Rule>> = ({ obj }) => {
-  const { t } = useTranslation(process.env.I18N_NAMESPACE);
-  const { perspective } = usePerspective();
-
-  const title: string = obj.annotations?.description || obj.annotations?.message;
-
-  return (
-    <>
-      <td className={tableRuleClasses[0]} title={title}>
-        <div className="co-resource-item">
-          <MonitoringResourceIcon resource={RuleResource} />
-          <Link to={getRuleUrl(perspective, obj)} className="co-resource-item__resource-name">
-            {obj.name}
-          </Link>
-        </div>
-      </td>
-      <td className={tableRuleClasses[1]} title={title}>
-        <Severity severity={obj.labels?.severity} />
-      </td>
-      <td className={tableRuleClasses[2]} title={title}>
-        {_.isEmpty(obj.alerts) ? '-' : <StateCounts alerts={obj.alerts} />}
-      </td>
-      <td className={tableRuleClasses[3]} title={title}>
-        {alertingRuleSource(obj) === AlertSource.User ? t('User') : t('Platform')}
-      </td>
-    </>
-  );
-};
-
-const RulesPage_: React.FC = () => {
-  const { t } = useTranslation(process.env.I18N_NAMESPACE);
-  const { alertsKey, silencesKey, rulesKey, perspective, defaultAlertTenant } = usePerspective();
-
-  const data: Rule[] = useSelector((state: MonitoringState) =>
-    getObserveState(perspective, state)?.get(rulesKey),
-  );
-  const { loaded = false, loadError }: Alerts = useSelector(
-    (state: MonitoringState) => getObserveState(perspective, state)?.get(alertsKey) || {},
-  );
-  const silencesLoadError = useSelector(
-    (state: MonitoringState) => getObserveState(perspective, state)?.get(silencesKey)?.loadError,
-  );
-
-  const ruleAdditionalSources = React.useMemo(
-    () => getAdditionalSources(data, alertingRuleSource),
-    [data],
-  );
-
-  const rowFilters: RowFilter[] = [
-    // TODO: The "name" filter doesn't really fit useListPageFilter's idea of a RowFilter, but
-    //       useListPageFilter doesn't yet provide a better way to add a filter like this
-    {
-      filter: (filter, rule: Rule) => fuzzyCaseInsensitive(filter.selected?.[0], rule.name),
-      filterGroupName: '',
-      items: [],
-      type: 'name',
-    } as RowFilter,
-    alertStateFilter(t),
-    severityRowFilter(t),
-    {
-      defaultSelected: defaultAlertTenant,
-      filter: (filter, rule: Rule) =>
-        filter.selected?.includes(alertingRuleSource(rule)) || _.isEmpty(filter.selected),
-      filterGroupName: t('Source'),
-      items: [
-        { id: AlertSource.Platform, title: t('Platform') },
-        { id: AlertSource.User, title: t('User') },
-        ...ruleAdditionalSources,
-      ],
-      reducer: alertingRuleSource,
-      type: 'alerting-rule-source',
-    },
-  ];
-
-  const [staticData, filteredData, onFilterChange] = useListPageFilter(data, rowFilters);
-
-  const columns = React.useMemo<TableColumn<Rule>[]>(
-    () => [
-      {
-        id: 'name',
-        props: { className: tableRuleClasses[0] },
-        sort: 'name',
-        title: t('Name'),
-        transforms: [sortable],
-      },
-      {
-        id: 'severity',
-        props: { className: tableRuleClasses[1] },
-        sort: (rules: Rule[], direction: 'asc' | 'desc') =>
-          _.orderBy(rules, alertSeverityOrder, [direction]) as Rule[],
-        title: t('Severity'),
-        transforms: [sortable],
-      },
-      {
-        id: 'state',
-        props: { className: tableRuleClasses[2] },
-        sort: (rules: Rule[], direction: 'asc' | 'desc') =>
-          _.orderBy(rules, alertingRuleStateOrder, [direction]),
-        title: t('Alert state'),
-        transforms: [sortable],
-      },
-      {
-        id: 'source',
-        props: { className: tableRuleClasses[3] },
-        sort: (rules: Rule[], direction: 'asc' | 'desc') =>
-          _.orderBy(rules, alertingRuleSource, [direction]),
-        title: t('Source'),
-        transforms: [sortable],
-      },
-    ],
-    [t],
-  );
-
-  return (
-    <>
-      <Helmet>
-        <title>Alerting</title>
-      </Helmet>
-      <div className="co-m-pane__body">
-        <ListPageFilter
-          data={staticData}
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore TODO
-          labelFilter="observe-rules"
-          labelPath="labels"
-          loaded={loaded}
-          onFilterChange={onFilterChange}
-          rowFilters={rowFilters}
-        />
-        {silencesLoadError && <SilencesNotLoadedWarning silencesLoadError={silencesLoadError} />}
-        <div className="row">
-          <div className="col-xs-12">
-            <VirtualizedTable<Rule>
-              aria-label={t('Alerting rules')}
-              label={t('Alerting rules')}
-              columns={columns}
-              data={filteredData ?? []}
-              loaded={loaded}
-              loadError={loadError}
-              Row={RuleTableRow}
-              unfilteredData={data}
-              NoDataEmptyMsg={() => {
-                return <EmptyBox label={t('Alerting rules')} />;
-              }}
-            />
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-const RulesPage = withFallback(RulesPage_);
+import AlertRulesPage from './alerting/AlertRulesPage';
 
 const Tab: React.FC<{ active: boolean; children: React.ReactNode }> = ({ active, children }) => (
   <li
@@ -333,7 +82,7 @@ const AlertingPage: React.FC<RouteComponentProps<{ url: string }>> = ({ match })
       </ul>
       <Switch>
         <Route path={alertsPath} exact component={AlertsPage} />
-        <Route path={rulesPath} exact component={RulesPage} />
+        <Route path={rulesPath} exact component={AlertRulesPage} />
         <Route path={silencesPath} exact component={SilencesPage} />
       </Switch>
     </>
